go-server/memory-rag/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
qdrant-client>=1.7.0
requests>=2.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
volcenginesdkarkruntime
